+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 ====================================================================
 ## 1.0.2
 * Fixed customizer select sanitize callback.
+* All text in theme dashboard are now internationalized.
 
 ## 1.0.1
 * Add more widgets and blocks.